SceneProcedural : Avoid computing attributes twice.
This shaves 9% off the time to first pixel for a complex benchmark scene.

Also removed the `full` argument to updateAttributes() to simplify the logic. The `full` argument was used to inherit the motion blur settings from parent locations if the root procedural was for a location other than the root of the scene. We weren't doing the same for the transform or attributes we output though, so removing `full` makes us more consistent there. In practice, the root procedural in a render always represents the root of the scene, so nothing has changed.