ValuePlug : Reject children which are not ValuePlugs.
Previously we accepted all children because the CompoundPlug derived class needed that for backwards compatibility. We now only provide the compatibility when we detect that we are actually a CompoundPlug. This closes a loophole that allowed abuse of ValuePlugs, and will allow us to detect any problems before we finally remove CompoundPlug.