ValuePlug bindings : Allow subclassing in Python.
At this stage the subclassing is only for ValuePlugs which intend to have children rather than for leaf plugs which intend to store values directly. This is equivalent to the old ability to subclass from CompoundPlug. It would be possible to add bindings to allow the creation of new leaf-level ValuePlugs but I'd like to encourage the use of C++ plugs as far as possible.