PathMatcher : Reference count internal nodes.
This does increase the size of a node by almost 30%, but it provides many opportunities for sharing nodes between PathMatchers in future, which we hope will result in a net reduction in memory usage. We may also be able to always reference the same node for the terminating leaf nodes, which alone may reduce memory usage below original levels.

In practice for a large production set, set memory usage increases by a more modest 11% overall, because much of the overhead is fixed in the storage of strings.