PathMatcher : Use a single shared node for all leaf nodes.
Since leaf nodes are often the most common within a tree, this gives us significant memory savings - typically enough to more than offset the increased size of a reference counted node over. Compared to master, this gives memory savings of 35% and 50% for the deep and shallow hierarchies constructed in PathMatcherTest.testLookupScaling (node overhead only, so in reality the savings are less because the fixed overhead of the InternedStrings remains the same).