Shader tests : Test dirty propagation through ShaderAssignment.
`ShaderAssignment::shaderPlug()` is always just a Plug with no children, but `Shader::outPlug()` may be a compound plug - for an Arnold shader for instance it will be a Color3fPlug. In this case we need to propagate dirtiness from the parent level of the Color3fPlug to the leaf level of the Plug.

This test passes currently - it is being added to ensure that the functionality remains when refactoring the dirty propagation mechanism.