Gaffer : Add DownstreamIterator.
This provides a straightforward way of doing a depth-first traversal through all the dependents of a particular plug. Until now this traversal logic was hidden away in the Plug dirty progation mechanism - this provides an official way of doing the same in the public API.

Note that this new mechanism differs from the original in that it uses the existence of children to decide if a plug is a leaf plug or not, whereas the former tested to see if a plug was a CompoundPlug. The CompoundPlug type is being phased out.