Plug : Handle exceptions during emission of dirtiness.
We don't want these to make their way out to the caller that triggered the dirty propagation. An error here generally means a bug in a `DependencyNode::affects()` implementation or a failure to prevent a cycle in `Plug::setInput()`, and there's nothing the caller can be expected to do about it.