NodeBinding : Optimise NodeWrapper::isInstanceOf( DependencyNode )
This query is especially prevalent because it is used by DownstreamIterator during dependency propagation and cycle prevention. In the case of a python class derived from NodeWrapper it is incredibly costly to enter python only to return false anyway. So we early-out in NodeWrapper, knowing that if we were actually a DependencyNode, the DependencyNodeWrapper override would correct our optimism.