GraphComponent : Scope dirty propagation during destruction.
The removal of child nodes and plugs inevitably results in a lot of dirty propagation activity, and there's no need to notify observers until it's all over.

This fixes several errors in InteractiveRenderManRenderTest, where the render was being updated halfway through destruction of the script, while the standard options were still specifying a camera, but after the camera had been destroyed.