Reduced hash cache clear frequency
It looks like computing hashes in the instance scene is quite a bottleneck when using an instancer node. It turns out you can reduce procedural expansion time in an instancing scene by a factor of 6 by turning down the hash cache clearing frequency, at a memory cost of about 100 mb, which seems acceptable. This change does not appear to affect performance on scenes with no instancer.