Isolate : Fix childNames hash.
We need to explicitly hash in the filter result for each of the child locations, rather than the result for _this_ location. Our hash must match the actions of our compute, and the compute uses the filter at the child locations.

There's no test case for this, because Gaffer itself doesn't have any filters sophisticated enough to trigger the bug. Instead I'm providing a separate test case in the IE internal project with the filter that exposed this bug.