GLWidget : Support use in Maya when using PySide.
The PySide bindings for QGLContext are incomplete compared to the PyQt ones, so we can't use the same trick to create a context to share with a host. Instead we reimplement our custom context in C++ so that it can be used regardless of the flavour of Qt bindings in use. As a bonus, we now have a destructor and can properly clean up our context.