RecursiveChildIterators : Fix API for completion of iteration.
This was vulnerable to the same bug documented in the previous DownstreamIterator fix. It seems less likely that it would manifest in this case, and indeed we've been aware of no such errors in years of use, but tracking down the first problem was hard enough that we should avoid doing it again in the future.

The previous API remains in deprecated form to allow a grace period for moving over client code.