DownstreamIterator : Fix API for completion of iteration.
The DownstreamIterator keeps a stack of Levels to represent recursion through the downstream plugs. Each Level stores a vector of plugs to visit, an iterator to the plug currently being visited, and an iterator to the end of the vector of plugs. The current position is therefore the iterator from the top of the stack, and the end is the end iterator for stack level 0. When current == end we're done right? Wrong. When stackSize > 1 that means we're comparing iterators from different containers, which is undefined. But still, vector::iterator is kinda like a pointer right, so we'll get away with it in practice? Wrong. In practice, the vector for one stack level might be allocated to reside in memory immediately after the vector for the previous stack level. Giving us the interesting situation where the end iterator for stack level 0 points to the same address as the start iterator for stack level 1.

Level   It            Start         End
======|=============|=============|=============
0       0x107b7fec0   0x107b7fec0   0x107b7fed0
1       0x107b7fed0   0x107b7fed0   0x107b7fed8

This leads to premature termination of the downstream iteration, and a test failure in DependencyNodeTest. The fix is to have an explicit `done()` method which also checks that the stack is of size 1.