RendererAlgo : Add outputTransform() method.
This deals with motion blur, and is smart enough to omit motion blocks when it detects static objects for which transform motion blur was requested. It uses an intermediate `transformSamples()` method to compute the samples - this will be useful in places like InteractiveRender where we wish to sample the scene on a different thread than the one the renderer API is used on.

Also, update SceneProcedural to use outputTransform().