SceneSwitch : Remove scene:path from context for index.
This makes it impossible to vary the index with scene location, which would allow the user to produce invalid scenes. This has the nice side effect of also reducing the number of unique contexts in which an index expression is evaluated,
which produces significant performance gains in a certain production scene.