FormatData : Make hash stable with GCC 4.4.7.
This one is a bit beyond my understanding. While working with the new PerformanceMonitor I noticed that an expression was being hashed more often than it should have been, and traced it down to the hash for the "image:defaultFormat" context variable changing each time it was computed.

This only occurred with GCC 4.4.7 - the new unit test already passed with GCC 4.8.3 and 4.6.3 and 5, and Clang 3.4. I got as far as verifying that the assembly for `SimpleDataHolder<GafferImage::Format>::hash()` was different in GCC 4.4.7, but at that point things were beyond me. The simple change to referencing rather than copying the Format object fixes things, and at this point I'm afraid to say that's good enough for me.