ScriptProcedural : Fix cache clearing when rendering in Arnold.
By connecting a member function to the allRendereredSignal we were relying on the renderer to keep the ScriptProcedural alive until all the child procedurals had been rendered. It appears that this is the case in 3delight (although we have no proof it will always be) but not in Arnold. This meant that `allRendered()` was never called in Arnold, and we never cleared the cache.

The fix is to connect a static method to allRenderedSignal, so we're not reliant on procedural lifetime at all. In general, we should make no assumptions about procedural lifetimes as this is in the hands of the renderer.