ScriptProcedural : Report errors via Node::errorSignal().
The child SceneProcedurals attempt to do error reporting via exception handling, but they're not in a position to know which upstream node caused the problem. This means the error messages they report lack context. By doing error reporting at a higher level via `Node::errorSignal()`, we have access to the source plug and can provide both the upstream plug name and the scene location that caused the problem.

This does mean that we are currently reporting errors twice, first in the ScriptProcedural and then in the SceneProcedural. We could remove the error reporting from SceneProcedural, except that it is used directly via AppleseedRender, so we would lose all reporting there. It seems that SceneProcedural may well be replaced with our own multithreaded scene expansion before too long, and in the meantime a belt and braces approach to error reporting shouldn't cause any real problems.