GraphGadget : Tighten rules for inserting nodes into noodles.
We now only insert if the node has a pair of in->out plugs where `in == DependencyNode::correspondingInput( out )`, and if those plugs can be inserted into the connection. When this is the case, it is also the case that disabling the inserted node creates a pass-through equivalent to the original connection, and  deleting the node will restore the original connection. This provides meaningful semantics - "insert node into stream". Previously we would insert nodes in such a way that one end of the connection was broken, or two new connections were made which had no logical in->out connection through the node.

In addition to more meaningful semantics, this will cut down on the unwanted noodle-snatching that has vexed the creators of particularly dense graphs.

This does mean that ExecutableNodes will no longer auto-insert, but that will be resolved when we rederive those from DependencyNode. Additionally, some RenderMan shaders which would auto-insert will no longer do so - either those insertions didn't make semantic sense, or the shaders should be updated to specify the "primaryInput" annotation, thereby defining the necessary `correspondingInput()` semantics.

Also :

- Use const where possible.
- Don't create inconsistent member state (setting only m_dragReconnectCandidate but not m_dragReconnect*Nodule) which has to be reverted on every early return from updateDragReconnectCandidate().