FileSystemPathTest : Prefer assertLess( a, b ) to assertTrue( a < b ).
These tests fail occasionally on Travis for unknown reasons, and this change should mean we get valuable information on the magnitude of the error. If it's small, we likely just have some latency on Travis and can up the threshold. If it's large, it's more likely that we have a timezone bug of some sort that needs fixing.