Add basic Sphinx/Recommonmark/ReadTheDocs documentation setup.
Sphinx is based on Python's docutils, which is beneficial to us because it comes from the same basic Python ecosystem we are already working in. This makes it easier to customise or fix things if necessary.

Sphinx uses RestructuredText as its default markup language. Although this is powerful, it's not as familiar or quite as readable as the Markdown syntax we use day-in day-out on GitHub. We use recommonmark to add a Markdown parser to the sphinx config, allowing us to parse Markdown files and integrate them into the documentation. The aim is to use Markdown for all the documentation, falling back to RestructuredText only where absolutely necessary.

Finally, ReadTheDocs provides a nice clean theme for presenting Sphinx documentation in HTML, and also has good online hosting. Here we're just using the theme - if we want to use the hosting we will need to set that up later.