SConstruct/config : Enable documentation build at Image Engine.
This was a little fiddly, because IE builds do not contain the dependencies internally as a standard build does, and we actually need to run the gaffer executable to generate the docs.

The SConstruct now has a LOCATE_DEPENDENCY_PYTHONPATH variable to specify the locations of python modules needed to run Gaffer, and the IE config points that at the necessary external modules. This is then used in the PYTHONPATH during the documentation build.

We no longer directly use `os.environ` to provide the build environment for documentation. Although this resulted in more setup work (HOME, DISPLAY and the Arnold/RenderMan paths), it's in keeping with the way SCons is intended to work.