UVWarp : Fix crashes triggered by negative data window origins.
We were relying on integer truncation to go from floating point pixel coordinates to integer coordinates. This is all well and good when the coordinates are positive :

    0.5, 0.5 -> 0, 0 (centre of pixel truncates to bottom left corner of pixel, aka its index)

But totally bogus when the coordinates are negative :

    -0.5, -0.5 -> 0, 0 (centre of pixel truncate to top right corner of pixel, aka index of a different pixel)

This was causing the UVWarp to index pixels outside of the bounds of it's tile buffers, and therefore crashing.