SystemCommand : Don't swallow stdout.
At some point we might want a mechanism the LocalDispatcher can use for redirecting this to the jobs window, but for now at least we'll be able to see it in the terminal. For dispatchers like TractorDispatcher, allowing the output direct to stdout is ideal anyway, because that's already the primary means of getting output into the logs.