SConstruct : Simplify packaging.
As noted in the \todo, gafferDependencies uses its own manifest already, so we only inherit the right files from that, and any files we add from the SConstruct here are added because we need them.

This fixes a packaging problem where libopenvdb was being excluded, because the gafferDependencies manifest was updated to include it, but the manifest here wasn't.