SConstruct : Convert BUILD_DIR to an absolute path.
This is motivated by a problem when building the documentation for the public releases, since the config/release/release.py script specifies BUILD_DIR as a relative path. This meant that we were also setting the PATH for the documentation build to contain a relative path, which broke when we changed directories while running the documentation auto-generation scripts. Since the PATH was now invalid, we were falling back to unwanted paths inherited from the environment, meaning we either built the docs for the wrong Gaffer version, or Gaffer wasn't found at all.