GafferUI : Add **kw arguments to all Widget constructors.
These are still needed to pass arguments to the base class constructors, and to allow the base class constructors to add arguments easily in the future. This is basically a revert of 6f0ee6c, but complicated slightly by the need to keep the Widget changes and a few scattered bug fixes intact.