Attributes : Avoid unnecessary dirty propagation.
Because the Attributes node has a mode where it affects attributes at locations, and a mode where it affects global attributes, it was unconditionally dirtying the output plugs for each mode when a setting changed. This led to unnecessary interactive render updates, which are best avoided.

The approach to fixing this is to manage pass-through connections so provide an optimised path where possible, and then to use the existence of the connection to disable dirty propagation. So this gives us a small pass-through optimisation as well as the larger optimisation of avoiding unneeded work in the InteractiveRender node.