TaskNode : Protect internal virtual methods.
The functionality provided by these should be accessed through the TaskPlug instead.

The friend TaskNodeAccessor is slightly awkward really. We could instead put its methods onto TaskNodeWrapper, if we were 100% sure that the only Python calls to TaskNode's protected methods were from reimplementations of those methods from derived classes. But since that functionality was only recently exposed publicly via TaskPlug and documented as being protected on TaskNode, I suspect that there will be client code depending on being able to make the old calls. If and when we come to do #1672, we could move the TaskNodeAccessor methods into TaskNodeWrapper, because we would know that they should only be called from derived python classes. For now though, we'll have to be content with just the improvements to the C++ interface.