Arnold Renderer : Don't apply default shader explicitly.
Arnold will automatically use a default shader if none is applied, and it behaves exactly as the one we were applying anyway. The key realisation to making this work is that `AiNodeSetPtr( ..., NULL )` is _not_ the same as `AiNodeResetParameter( ... )` - the former gives the "bad shader" error colour and the latter reverts to the default shading behaviour we want.

Not only is this simpler, it also means we no longer assign a default shader to external procedurals, where Arnold's odd forced inheritance behaviour would override all shaders on nodes within the procedural.