GafferOSL : Update for OSL 1.7.
Note that this is also dropping support for prior versions. We could keep support for older versions if it's really important to anyone, but we're about to start taking advantage of features of 1.7 (specifically varying length array parameters to support ramps as we do for RSL), and I'd prefer to have all Gaffer builds provide the same feature set rather than have some features missing. OSL 1.7 also changed closure registration a fair bit, and I'd prefer not to manage all the necessary #ifdeffing unless it's going to be of a genuine benefit somewhere.

The expectedFailure removed from the unit tests was presumably a genuine failure in some previous version of OSL, but it seems that now we're actually reading the metadata just fine, and it was an error in the test itself that we were catching as the expected failure.