Light nodes : Use IECore::Shader instead of IECore::Light.
Our use of IECore::Light has always been awkward, because it has a light handle property which we carry pointlessly through the scene, only giving it a real value at the point of output to the renderer. The only reason IECore::Light really existed was as an object to map to a call to IECore::Renderer::light() but even that purpose is unnecessary with the new renderer backend that we're introducing.

We still support the use of IECore::Light by derived classes that might live outside Gaffer, but emit a deprecation warning if one is found.