GafferArnold : Move path setup from __init__.py to wrapper.
This means that `gaffer env kick` can be used successfully to render a pregenerated ass file containing a Gaffer output driver, and also that a SystemCommand can be used to call `kick` even if no GafferArnold nodes are in the script.

Deliberately omitted the procedurals location, because we no longer use the Cortex python procedural for Arnold rendering.