Renderer : Require attributes to be passed at object creation time.
This is motivated by a need to support attributes that affect Arnold's tesselation settings on polymesh objects, as well as a desire to reintroduce automatic instancing capabilities into the renderer backends. In an ideal world, attributes and objects would be truly orthogonal, and the unique hash for an object undergoing automatic instancing would simply be the hash of the object. For the Arnold backend, that is unfortunately not the case since various attributes will need to be taken into account in generating the instancing hash.

This implies that attributes are needed at the point the object is converted to the renderer, presenting us with several implementation options :

1. Require an initial set of attributes to be passed to `Renderer::object()`, with only subsequent edits using `ObjectInterface::attributes()`.
2. Buffer the IECore::Object internally in the renderer, until the attributes are provided by `ObjectInterface::attributes()` and it can be converted. Mandate that client code _must_ call `Object::Interface::attributes()` so that we have this opportunity.
3. Add an `ObjectInterface::flush()` method which makes the synchronisation step from 2) more explicit.

Since it's impossible to enforce correct client behaviour for 2) and 3), we have implemented 1) here in this commit. It is a pity that initial attribute specification is not done through the same method as subsequent edits, but in practice (see unit tests) it's actually pretty convenient to be able to specify the attributes inline with the object creation.