Arnold Renderer : Add automatic instancing and subdivision attributes.
Where two or more objects are identical (share the same hash), we now automatically create them as instances of the same underlying node in Arnold. The only exception to this is where the subdivision settings for a polymesh are view dependent, in which case we cannot expect a single master mesh to be subdivided appropriately for all instances - in this case we automatically disable the instancing.