GafferScene : Add ShaderPlug.
This continues the trend of introducing new plug types to enforce strong typing where before we used `Node::acceptsInput()` overrides as a crutch. Benefits are clearer semantics, centralised logic for connection acceptance, and the ability to create new node types which have an external ShaderPlug connected into an internal ShaderAssignment.