OSLExpressionEngine : Fix string identity bug.
OSL seems to store all strings as ustring, and == just compares addresses, so we need to uniquefy our string values with `ustring()` rather than `InternedString`. I'm not 100% sure if the `*(const char **)` cast is right, or if it should be `*(ustring *)`, but since it has appeared to work until now, I'm not changing it without further evidence.

This fixes one of the string comparison bugs reported in #1789.