GafferScene : Add FilterPlug.
To date, we've been using a plain old IntPlug to provide the output of a Filter node and the input to a FilteredSceneProcessor. But this alone was insufficient to provide the guarantee that only Filter outputs could be connected to a FilteredSceneProcessor input. We used FilteredSceneProcessor::acceptsInput() to add constraints in an attempt to enforce this, but these have their problems as witnessed by #1814 and `SceneProcessorTest.testDeriveAndUseExternalFilters`.

The FilterPlug is intended to provide a stronger mechanism for making the guarantees we need. By using a dedicated plug type, we ensure that the plug semantics are preserved when undergoing promotion, and nodes can rely on the plug type rather than a custom `acceptsInput()` method which will become more flawed as other means of duplicating/promoting plugs appear in the future.

This is a pattern that has repeated itself every single time we've used a less strongly typed plug along with a custom `Node::acceptsInput()` implementation, rather than use a custom plug type from the outset. We've already made an analogous transition from Plug to TaskPlug in GafferDispatch, and shortly will be adding a ShaderPlug for use as the input to a ShaderAssignment. The rationale for the original implementations was to reduce the proliferation of many specialist plug types, but it seems clear at this point that the general rule should be that a custom plug type should be preferred to `Node::acceptsInput` overloads.