Arnold ParameterHandler : Support gaffer.plugType metadata.
Currently only implemented for RGBA parameters, where you can now optionally represent them as Color3fPlugs in Gaffer. Rewrote the plug creation for RGB and RGBA parameters to allow the reuse of suitable plugs in the same way as we do for `OSLShader::loadShader()`, making some progress towards adding a `keepExistingValues` argument to `ArnoldShader::loadShader()`.

The main use of this is to use a metadata file to cause ray_switch and bump shaders to have Color3f inputs compatible with the outputs from most regular shaders. This is something of a stopgap measure until we can improve the UI to support RGB->RGBA connections well.

Requires ImageEngine/cortex#514 to be useful, but doesn't have a hard dependency.

Breaking Changes :

- Added an argument to GafferArnold::ParameterAlgo::setupPlug().