ObjectSource, Light: change interface for default set membership
The idea here is to provide an interface that allows an object to be
part of multiple sets per default. We use this at IE to add lights to
both "__lights" and a set based on the light type.

John, instead of adding that additional plug that we talked about, this
will work without that extra level of indirection. How do you feel about
this solution? I made the corresponding changes in IERendering
already and that seems to work fine.