GadgetWidget : Manage ViewportGadget visibility.
We set the visibility of the ViewportGadget based on the visibility of the containing GadgetWidget, so that Gadget-level code can be informed of the overall visibility of the gadgets on screen via the new `Gadget::visibilityChangedSignal()`.

Also added a GIL release in the `Gadget::setVisible()` binding, because the `visibilityChangedSignal()` code path now allows arbitrary code to be executed, which may need to access python on another thread.