CompoundNodule : Avoid creation of duplicate child nodules.
This occurred when a CompoundNodule was constructed from a `childAddedSignal()`, via the OSLCode node changing it's metadata. In this case the CompoundNodule would create all the children it needed in the constructor, connect to `childAddedSignal()`, and then create a duplicate as `childAdded()` was called from the (already in flight) signal.