StringPlug : Fix substitutions when plug has an input.
We perform substitutions automatically in getValue() when called on an input plug within a compute. This provides necessary features to the user while freeing the developer from the burden of having to think about substitutions. When computing and hashing a plug, we first travel of the chain of inputs (if there is one), and do the compute at the source, to avoid the cost of repeating a series of intermediate computations and inserting them in the cache - this is an important optimisation when many nodes contain pass-throughs internally.

We had a bug in the interaction between these two things, whereby we were determining whether substitutions should occur in `getValue()` using the direction of the plug itself, and whether they should be accounted for in `hash()` using the direction of the source plug. When the source was an output, this led to a non-time-varying hash, which meant that results were reused from the cache inappropriately.

There are two fixes here. Firstly, we make sure the `hash()` and `getValue()` methods perform substitutions identically. Secondly, we also take account of substitution flags on all intermediate plugs when a chain of connections leads to an input. This resolves the "which plug flags are relevant?" question, by taking them all into account, rather than only those at the end of the chain. This means the "go straight to the source" optimisation now acts as if the value was passed along by a series of `getValue()` calls on the intermediate plugs - that is, as if the optimisation did not exist.

This will be visible to users as follows :

- When driving a plug with an expression outputting strings containing '#' or '${variable}' substitutions, the downstream node will now compute correctly in all contexts, instead of appearing frozen on the results of the first compute.
- If a plug with substitutions turned off (ImageReader fileName, or FileSequenceParameter for instance) receives its value from an intermediate plug with substitutions turned on, it will not receive substitutions. If the intermediate plug was created by promotion, it will have the appropriate flags anyway, and act is before, but if created by other means, the behaviour may change.

I'm concerned about the latter change. In practice, substitution settings are largely determined based on the needs of the node owning the plug, so having it react different based on where the input comes from seems inappropriate. Stand by...