ArnoldShaderUI : Support layout section metadata.
We support both OSL style "page" metadata, and also translate the baffling metadata format used in Houdini, because that means we can make decent layouts for AlShaders. HtoA also ships with layout metadata for the standard Arnold shaders, so if that is configured on the ARNOLD_PLUGIN_PATH we get better layout for the standard shaders too.