AppleseedShaderAdaptor : Omit input shaders if not needed.
When we're just outputting the red error shader, we don't need to keep the original shaders at all. We should probably be omitting them for the sake of tidiness and clarity anyway, but right now we're further motivated by the fact that Appleseed appears to crash when processing shaders with string outputs, so this commit at least prevents the shader ball from crashing when previewing a shader with a string output.