Arnold Renderer : Add log verbosity options.
We expose each individual message flag as a boolean option rather than expose a single bitmask for everything. This is in keeping with the treatment for the visibility flags. An alternative would be to synthesize a single integer "verbosity level" option similar to the one provided in `kick`. The kick one takes on values 0-6, with higher values turning on progressively more flags. That approach uses 7 distinct values which is less than the 13 used here, but is somewhat opaque in that its not clear what is turned on by what value. So I've gone with dual "expose the renderer features directly" and "explicit trumps implicit" justifications for exposing each flag individually.