SceneInspector : Fix errors when selected path doesn't exist.
When comparing two scenes, the selected path might not exist in both scenes. In this case, the object section was trying to generate an object at the invalid path, and causing errors. This is avoided by checking the path is not none, as this is the signifier for an invalid path.