Serialiser : Give serialisation argument to all methods.
We started out with a serialisation argument on just a couple of the methods, then added one to the `constructor()` method. We now need the serialisation in the `childNeedsConstruction()` method, so might as well add it to the rest too on the grounds that it might reasonably be needed in the future.

This commit only affects function signatures - no behaviour is changed.

Breaking change :

- Changed signatures of Serialiser virtual methods.