Selection/Crop Tools : Fix precision issues when far from origin.
The p0 of our line is on the near clipping plane. We get better precision by using the far value from p1 instead.

Another way of approaching this might be to improve the API so the gadgets aren't forced to project back into raster space just to get the position of an event which originated in raster space anyway. It has always been a bit awkward that ButtonEvent contains only a `LineSegment3f line` field, when Widgets only care about 2 dimensions, and many Gadgets also care about the original raster position. Maybe we should add a `V2f point` field to ButtonEvent, which holds the 2d Widget-relative position for Widgets, and can then be used by Gadgets for retrieving the original and exact raster position.