NodeGraph : Navigate into Reference nodes with Down key.
We used to require the use of a "secret" set of modifiers for this, because the internals were not read-only as they should have been, and making apparently successful edits which were lost on serialisation was confusing. In practice though, everyone discovered the secret modifiers and got confused anyway. Some folks even thought you needed the full set of modifiers just to navigate into a regular box.

We still allow navigation into the internals of other node types for debugging, using the same old modifiers. But we no longer allow navigation into Boxes or References when the modifiers are used, to teach folks that they are not the standard way of doing things.