Metadata : Remove `inherit` arguments from API.
These existed only for use in testing that nodes were documented, as per #232. They otherwise cluttered up the API, and made querying instanceOnly metadata from C++ really confusing - it was necessary to pass a value for `inherit` even though it would be irrelevant.

The old deprecated methods keep their arguments intact, because we're only keeping those around for temporary backwards compatibility anyway. The `plugsWithMetadata()` and `nodesWithMetadata()` are changed because they are part of the standard API - this is a compatibility break, but no IE code (at least on GitHub) currently uses that argument.

The documentation tests are now performed by querying the base class for each node using Python's introspection. This is a bit fiddlier, but better to have a minor complication in the implementation of one method than to clutter the public API for all users.

Breaking changes :

- Removed `inherit` argument from `Metadata.plugsWithMetadata()`
- Removed `inherit` argument from `Metadata.nodesWithMetadata()`