SceneInspector.SideBySideDiff : Support overrides for visibilities
The history and inheritance sections were doing a bit of naughty after-the-fact management of the visibilities that the SideBySideDiff was setting internally. Now they have a legitimate mechanism for getting it right in the first place. This change will allow us to better encapsulate the SideBySideDiff, removing the public visibility of the frames.