Wrapper : Avoid duplicate paths.
When launching a Gaffer process as a child of another Gaffer process (for instance, when doing a dispatch), the wrapper was adding duplicate paths. We now only add them if they are not there. This generalises the previous Appleseed commit nicely, so that we fix both the SIP issues as well as the duplication of path values, and also simplify the logic and readability.