Arnold Renderer : Fix "ai:shape:step_size" support.
This was broken when we introduced the new renderer backend. We also now support the application of step_size to mesh geometries, in which case we convert the geometry to an Arnold box node sized to match the bounding box of the mesh. Arnold doesn't yet support rendering meshes as volumes, so this is the best we can really be expected to do, and is the same approach as taken by MtoA.