StandardNodule : Support "label" metadata.
If labels were typically visible all the time, we would probably want to cache the metadata lookup in a member variable to avoid doing the lookup in `renderLabel()`. But in practice very few labels are visible at any given time, and doing the lookup in `renderLabel()` allows us to maintain binary compatibility.